+++ resources/scripts/providers/index.js
@@ -10,5 +10,3 @@ const apexFormatterProvider = require('./apexFormatterProvider');
 exports.apexFormatterProvider = apexFormatterProvider.provider;
 const fileStructureProvider = require('./fileStructureProvider');
 exports.FileStructureProvider = fileStructureProvider.FileStructureTreeProvider;
-const onTypeFormattingProvider = require('./onTypeFormattingProvider');
-exports.onTypeFormattingProvider = onTypeFormattingProvider.provider;