+++ resources/scripts/providers/javascriptCompletionProvider.js
@@ -10,6 +10,7 @@ const CompletionItemKind = vscode.CompletionItemKind;
 const CompletionItem = vscode.CompletionItem;
 const MarkdownString = vscode.MarkdownString;
 const SnippetString = vscode.SnippetString;
+const langUtils = languages.Utils;
 
 exports.provider = {
     provideCompletionItems(document, position) {
@@ -23,14 +24,25 @@ exports.provider = {
 
 function provideJSCompletion(document, position) {
     let items;
-    let activation = Utils.getActivation(document, position);
+    let activationInfo = Utils.getActivation(document, position);
+    let activation = activationInfo.activation;
     let activationTokens = activation.split('.');
     let similarJSSnippetsNs = getSimilarSnippetsNS(applicationContext.jsSnippets, activationTokens[0]);
     let snippets;
+    let queryData = langUtils.getQueryData(document, position);
+    let classes = Utils.getClassesFromClassFolder(document);
+    let systemMetadata = Utils.getNamespaceMetadataFile('System');
+    let namespacesMetadata = Utils.getNamespacesMetadataFile();
+    let sObjects = Utils.getObjectsFromMetadataIndex();
+    let componentStructure = BundleAnalizer.getComponentStructure(document.fileName.replace('Controller.js', '.cmp').replace('Helper.js', '.cmp'));
     if (FileChecker.isJavaScript(document.uri.fsPath) && applicationContext.jsSnippets[activationTokens[0]] && applicationContext.jsSnippets[activationTokens[0]].length > 0) {
         snippets = applicationContext.jsSnippets[activationTokens[0]];
     }
-    if (snippets && snippets.length > 0) {
+
+    if (queryData) {
+        // Code for support completion on queries
+        items = Utils.getQueryCompletionItems(activationTokens, activationInfo, queryData, position, 'aurahelper.completion.aura');
+    } else if (snippets && snippets.length > 0) {
         // Code for completions when user types any snippets activation preffix (ltn., slds., ltng. ...)
         items = getSnippetsCompletionItems(position, snippets);
     } else if (similarJSSnippetsNs.length > 0 && FileChecker.isJavaScript(document.uri.fsPath)) {
@@ -38,12 +50,16 @@ function provideJSCompletion(document, position) {
         items = getSimilarCompletionItems(position, similarJSSnippetsNs);
     } else if (activationTokens[0] === 'v' || activationTokens[0] === 'c' || activationTokens[0] === 'helper') {
         // Code for completions when user types v. c. or helper.
-        let componentStructure = BundleAnalizer.getComponentStructure(document.fileName.replace('Controller.js', '.cmp').replace('Helper.js', '.cmp'));
         if (activationTokens[0] === 'v') {
             // Code for completions when user types v.
             if (!config.getConfig().activeAttributeSuggest)
                 return Promise.resolve(undefined);
-            items = getAttributes(componentStructure, position, undefined);
+            let attribute = Utils.getAttribute(componentStructure, activationTokens[1]);
+            if (attribute) {
+                items = getComponentAttributeMembersCompletionItems(attribute, activationTokens, activationInfo, sObjects, position);
+            } else if (activationTokens.length === 2) {
+                items = getAttributes(componentStructure, position,);
+            }
         } else if (activationTokens[0] === 'c') {
             // Code for completions when user types c.
             if (!config.getConfig().activeControllerMethodsSuggest)
@@ -57,6 +73,302 @@ function provideJSCompletion(document, position) {
         }
     } else if (activationTokens.length > 0 && activationTokens[0].toLowerCase() === 'label') {
         items = getLabelsCompletionItems(activationTokens, position);
+    } else if (activationTokens.length > 1) {
+        // Code for completions when position is on empty line or withot components
+        items = getApexCompletionItems(document, position, activationTokens, activationInfo, classes, systemMetadata, namespacesMetadata, sObjects);
+
+    } else if (activationTokens.length > 0) {
+        // Code for completions when position is on empty line or withot components
+        items = getAllAvailableCompletionItems(position, classes, systemMetadata, namespacesMetadata, sObjects);
+
+    }
+    return items;
+}
+
+function getComponentAttributeMembersCompletionItems(attribute, activationTokens, activationInfo, sObjects, position) {
+    let items;
+    if (sObjects.sObjectsToLower.includes(attribute.type.toLowerCase())) {
+        if (!config.getConfig().activeSobjectFieldsSuggestion)
+            return Promise.resolve(undefined);
+        let sObject = Utils.getObjectFromMetadataIndex(sObjects.sObjectsMap[attribute.type.toLowerCase()]);
+        if (activationTokens.length >= 2) {
+            let lastObject = sObject;
+            let index = 0;
+            for (const activationToken of activationTokens) {
+                let actToken = activationToken;
+                if (index > 1) {
+                    if (actToken.endsWith('__r'))
+                        actToken = actToken.substring(0, actToken.length - 3) + '__c';
+                    let fielData = Utils.getFieldData(lastObject, actToken);
+                    if (fielData) {
+                        if (fielData.referenceTo.length === 1) {
+                            lastObject = Utils.getObjectFromMetadataIndex(fielData.referenceTo[0]);
+                        } else {
+                            lastObject = undefined;
+                        }
+                    }
+                }
+                index++;
+            }
+            items = Utils.getSobjectsFieldsCompletionItems(position, lastObject, 'aurahelper.completion.aura', activationTokens, activationInfo);
+        }
+    } else {
+        // include Apex Classes Completion
+    }
+    return items;
+}
+
+function getAllAvailableCompletionItems(position, classes, systemMetadata, namespacesMetadata, sObjects) {
+    let items = [];
+    for (const userClass of classes.classes) {
+        let options = Utils.getCompletionItemOptions(userClass, 'Custom Apex Class', userClass, true, CompletionItemKind.Class);
+        let command = Utils.getCommand('UserClass', 'aurahelper.completion.aura', [position, 'UserClass', userClass]);
+        items.push(Utils.createItemForCompletion(userClass, options, command));
+    }
+    Object.keys(systemMetadata).forEach(function (key) {
+        let systemClass = systemMetadata[key];
+        if (systemClass.isEnum) {
+            let description = systemClass.description + ((systemClass.link) ? 'Documentation:\n ' + systemClass.link : '') + '\nEnum Values: \n' + systemClass.enumValues.join('\n');
+            let options = Utils.getCompletionItemOptions('Enum from ' + systemClass.namespace + ' Namespace', description, systemClass.name, true, CompletionItemKind.Enum);
+            let command = Utils.getCommand('SystemEnum', 'aurahelper.completion.aura', [position, 'SystemEnum', systemClass]);
+            items.push(Utils.createItemForCompletion(systemClass.name, options, command));
+        } else if (systemClass.isInterface) {
+            let description = systemClass.description + ((systemClass.link) ? 'Documentation:\n ' + systemClass.link : '');
+            let options = Utils.getCompletionItemOptions('Interface from ' + systemClass.namespace + ' Namespace', description, systemClass.name, true, CompletionItemKind.Interface);
+            let command = Utils.getCommand('SystemInterface', 'aurahelper.completion.aura', [position, 'SystemInterface', systemClass]);
+            items.push(Utils.createItemForCompletion(systemClass.name, options, command));
+        } else {
+            let description = systemClass.description + ((systemClass.link) ? 'Documentation:\n ' + systemClass.link : '');
+            let options = Utils.getCompletionItemOptions('Class from ' + systemClass.namespace + ' Namespace', description, systemClass.name, true, CompletionItemKind.Class);
+            let command = Utils.getCommand('SystemClass', 'aurahelper.completion.aura', [position, 'SystemClass', systemClass]);
+            items.push(Utils.createItemForCompletion(systemClass.name, options, command));
+        }
+    });
+    Object.keys(namespacesMetadata).forEach(function (key) {
+        let nsMetadata = namespacesMetadata[key];
+        let options = Utils.getCompletionItemOptions(nsMetadata.description, nsMetadata.docLink, nsMetadata.name, true, CompletionItemKind.Module);
+        let command = Utils.getCommand('Namespace', 'aurahelper.completion.aura', [position, 'Namespace', nsMetadata]);
+        items.push(Utils.createItemForCompletion(nsMetadata.name, options, command));
+    });
+    if (config.getConfig().activeSObjectSuggestion) {
+        for (const sobject of sObjects.sObjectsToLower) {
+            let objName = sObjects.sObjectsMap[sobject];
+            let splits = objName.split('__');
+            let namespace = '';
+            let description = 'Standard SObject';
+            if (objName.indexOf('__c') !== -1)
+                description = 'Custom SObject';
+            if (splits.length > 2) {
+                namespace = splits[0].trim();
+                description += '\nNamespace: ' + namespace;
+            }
+            let options = Utils.getCompletionItemOptions(objName, description, objName, true, CompletionItemKind.Class);
+            let command = Utils.getCommand('SObject', 'aurahelper.completion.aura', [position, 'SObject', objName]);
+            items.push(Utils.createItemForCompletion(objName, options, command));
+        }
+    }
+    return items;
+}
+
+function getApexCompletionItems(document, position, activationTokens, activationInfo, classes, systemMetadata, namespacesMetadata, sObjects) {
+    let items = [];
+    let sObject = Utils.getObjectFromMetadataIndex(sObjects.sObjectsMap[activationTokens[0].toLowerCase()]);
+    let lastClass = undefined;
+    let parentStruct;
+    let index = 0;
+    for (let actToken of activationTokens) {
+        if (index < activationTokens.length - 1) {
+            let actType = Utils.getActivationType(actToken);
+            let datatype;
+            let className;
+            if (sObject) {
+                if (actToken.endsWith('__r'))
+                    actToken = actToken.substring(0, actToken.length - 3) + '__c';
+                let fielData = Utils.getFieldData(sObject, actToken);
+                if (fielData) {
+                    if (fielData.referenceTo.length === 1) {
+                        sObject = Utils.getObjectFromMetadataIndex(fielData.referenceTo[0]);
+                    } else {
+                        datatype = fielData.type;
+                        if (datatype.indexOf('<') !== -1)
+                            datatype = datatype.split('<')[0];
+                        if (datatype.indexOf('[') !== -1 && datatype.indexOf(']') !== -1)
+                            datatype = "List";
+                        if (datatype.indexOf('.') !== -1) {
+                            let splits = datatype.split('.');
+                            if (splits.length === 2) {
+                                let parentClassOrNs = splits[0];
+                                className = splits[1];
+                                if (systemMetadata[parentClassOrNs]) {
+                                    parentStruct = Utils.getClassStructure(document, 'System', parentClassOrNs);
+                                } else if (namespacesMetadata[parentClassOrNs]) {
+                                    lastClass = Utils.getClassStructure(document, parentClassOrNs, className);
+                                    parentStruct = undefined;
+                                    sObject = undefined;
+                                }
+                            } else if (splits.length === 2) {
+                                let nsName = splits[0];
+                                let parentClassName = splits[1];
+                                className = splits[2];
+                                if (systemMetadata[parentClassName.toLowerCase()]) {
+                                    parentStruct = Utils.getClassStructure(document, 'System', parentClassName);
+                                } else if (namespacesMetadata[nsName.toLowerCase()]) {
+                                    lastClass = Utils.getClassStructure(document, nsName, parentClassName);
+                                    parentStruct = undefined;
+                                    sObject = undefined;
+                                }
+                            }
+                        } else {
+                            parentStruct = undefined;
+                            if (systemMetadata[datatype.toLowerCase()]) {
+                                lastClass = Utils.getClassStructure(document, 'System', datatype);
+                                sObject = undefined;
+                            }
+                        }
+                        if (parentStruct && className) {
+                            let classFound = false;
+                            Object.keys(parentStruct.classes).forEach(function (key) {
+                                let innerClass = parentStruct.classes[key];
+                                if (innerClass.name.toLowerCase() === className.toLowerCase()) {
+                                    classFound = true;
+                                    lastClass = innerClass;
+                                }
+                            });
+                            if (!classFound) {
+                                Object.keys(parentStruct.enums).forEach(function (key) {
+                                    let innerEnum = parentStruct.enums[key];
+                                    if (innerEnum.name.toLowerCase() === className.toLowerCase()) {
+                                        lastClass = innerEnum;
+                                        classFound = true;
+                                    }
+                                });
+                            }
+                            if (!classFound)
+                                lastClass = undefined;
+                        }
+                    }
+                }
+            } else if (lastClass) {
+                if (!lastClass.isEnum) {
+                    if (actType.type === 'field') {
+                        if (lastClass.posData && lastClass.posData.isOnMethod) {
+                            let method = Utils.getMethod(lastClass, lastClass.posData.methodSignature);
+                            let methodVar = Utils.getVariable(method, actToken);
+                            let classVar = Utils.getClassField(lastClass, actToken);
+                            if (methodVar)
+                                datatype = methodVar.datatype;
+                            else if (classVar)
+                                datatype = classVar.datatype;
+                        } else {
+                            let classVar = Utils.getClassField(lastClass, actToken);
+                            if (classVar)
+                                datatype = classVar.datatype;
+                        }
+                    } else if (actType.type === 'method') {
+                        let method = Utils.getMethodFromCall(lastClass, actType.name, actType.params);
+                        if (method)
+                            datatype = method.datatype;
+                    }
+                    if (!datatype) {
+                        if (lastClass.parentClass) {
+                            parentStruct = Utils.getClassStructure(document, undefined, lastClass.parentClass);
+                            className = actToken;
+                        } else {
+                            if (classes.classesToLower.includes(actToken.toLowerCase())) {
+                                lastClass = Utils.getClassStructure(document, undefined, actToken);
+                                parentStruct = undefined;
+                                sObject = undefined;
+                            } else if (systemMetadata[actToken.toLowerCase()]) {
+                                lastClass = Utils.getClassStructure(document, 'System', actToken);
+                                parentStruct = undefined;
+                                sObject = undefined;
+                            } else if (sObjects.sObjectsToLower.includes(actToken.toLowerCase())) {
+                                sObject = Utils.getObjectFromMetadataIndex(actToken);
+                                parentStruct = undefined;
+                                lastClass = undefined;
+                            }
+                        }
+                    } else {
+                        if (datatype.indexOf('<') !== -1)
+                            datatype = datatype.split('<')[0];
+                        if (datatype.indexOf('[') !== -1 && datatype.indexOf(']') !== -1)
+                            datatype = "List";
+                        if (datatype.indexOf('.') !== -1) {
+                            let splits = datatype.split('.');
+                            if (splits.length === 2) {
+                                let parentClassOrNs = splits[0];
+                                className = splits[1];
+                                if (classes.classesToLower.includes(parentClassOrNs.toLowerCase())) {
+                                    parentStruct = Utils.getClassStructure(document, undefined, parentClassOrNs);
+                                } else if (systemMetadata[parentClassOrNs]) {
+                                    parentStruct = Utils.getClassStructure(document, 'System', parentClassOrNs);
+                                } else if (namespacesMetadata[parentClassOrNs]) {
+                                    lastClass = Utils.getClassStructure(document, parentClassOrNs, className);
+                                    parentStruct = undefined;
+                                    sObject = undefined;
+                                }
+                            } else if (splits.length === 2) {
+                                let nsName = splits[0];
+                                let parentClassName = splits[1];
+                                className = splits[2];
+                                if (classes.classesToLower.includes(parentClassName.toLowerCase())) {
+                                    parentStruct = Utils.getClassStructure(document, undefined, parentClassName);
+                                } else if (systemMetadata[parentClassName.toLowerCase()]) {
+                                    parentStruct = Utils.getClassStructure(document, 'System', parentClassName);
+                                } else if (namespacesMetadata[nsName.toLowerCase()]) {
+                                    lastClass = Utils.getClassStructure(document, nsName, parentClassName);
+                                    parentStruct = undefined;
+                                    sObject = undefined;
+                                }
+                            }
+                        } else {
+                            parentStruct = undefined;
+                            if (lastClass.parentClass && datatype !== 'List') {
+                                parentStruct = Utils.getClassStructure(document, undefined, lastClass.parentClass);
+                                className = datatype;
+                            } else if (classes.classesToLower.includes(datatype.toLowerCase())) {
+                                lastClass = Utils.getClassStructure(document, undefined, datatype);
+                                sObject = undefined;
+                            } else if (systemMetadata[datatype.toLowerCase()]) {
+                                lastClass = Utils.getClassStructure(document, 'System', datatype);
+                                sObject = undefined;
+                            } else if (sObjects.sObjectsToLower.includes(datatype.toLowerCase())) {
+                                sObject = Utils.getObjectFromMetadataIndex(datatype);
+                                parentStruct = undefined;
+                                lastClass = undefined;
+                            }
+                        }
+                    }
+                    if (parentStruct && className) {
+                        let classFound = false;
+                        Object.keys(parentStruct.classes).forEach(function (key) {
+                            let innerClass = parentStruct.classes[key];
+                            if (innerClass.name.toLowerCase() === className.toLowerCase()) {
+                                classFound = true;
+                                lastClass = innerClass;
+                            }
+                        });
+                        if (!classFound) {
+                            Object.keys(parentStruct.enums).forEach(function (key) {
+                                let innerEnum = parentStruct.enums[key];
+                                if (innerEnum.name.toLowerCase() === className.toLowerCase()) {
+                                    lastClass = innerEnum;
+                                    classFound = true;
+                                }
+                            });
+                        }
+                        if (!classFound)
+                            lastClass = undefined;
+                    }
+                }
+            }
+        }
+        index++;
+    }
+    if (lastClass && config.getConfig().activeApexSuggestion) {
+        items = Utils.getApexClassCompletionItems(position, lastClass);
+    } else if (sObject && config.getConfig().activeSobjectFieldsSuggestion) {
+        items = Utils.getSobjectsFieldsCompletionItems(position, sObject, 'aurahelper.completion.aura', activationTokens, activationInfo);
     }
     return items;
 }